requirements.txt
alembic==1.14.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.3
djhtml==3.0.7
djlint==1.36.4
EditorConfig==0.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsbeautifier==1.15.3
json5==0.10.0
libsass==0.23.0
Mako==1.3.9
MarkupSafe==3.0.2
pathspec==0.12.1
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
SQLAlchemy==2.0.37
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.2.1
